opendit update
