+++ codes/OpenDiT/README.md
@@ -6,9 +6,14 @@
 <p align="center"><a href="https://github.com/NUS-HPC-AI-Lab/OpenDiT">[Homepage]</a> | <a href="https://discord.gg/yXF4n8Et">[Discord]</a> | <a href="./figure/wechat.jpg">[WeChat]</a> | <a href="https://twitter.com/YangYou1991/status/1762447718105170185">[Twitter]</a> | <a href="https://zhuanlan.zhihu.com/p/684457582">[Zhihu]</a> | <a href="https://mp.weixin.qq.com/s/IBb9vlo8hfYKrj9ztxkhjg">[Media]</a></p>
 </p>
 
+###  Latest News 🔥
+
+* [2024/03/01] Support DiT-based Latte for text-to-video generation.
+* [2024/02/27] Officially release OpenDiT: An Easy, Fast and Memory-Efficent System for DiT Training and Inference.
+
 # About
 
-OpenDiT is an open-source project that provides a high-performance implementation of Diffusion Transformer(DiT) powered by Colossal-AI, specifically designed to enhance the efficiency of training and inference for DiT applications, including text-to-video generation and text-to-image generation.
+OpenDiT is an open-source project that provides a high-performance implementation of Diffusion Transformer (DiT) powered by Colossal-AI, specifically designed to enhance the efficiency of training and inference for DiT applications, including text-to-video generation and text-to-image generation.
 
 OpenDiT boasts the performance by the following techniques:
 
@@ -87,7 +92,7 @@ pip install -v --disable-pip-version-check --no-cache-dir --no-build-isolation -
 
 ### Image
 
-<b>Training.</b> You can train the DiT model by executing the following command:
+<b>Training.</b> You can train the DiT model on CIFAR10 by executing the following command:
 
 ```shell
 # Use script
@@ -95,29 +100,34 @@ bash train_img.sh
 # Use command line
 torchrun --standalone --nproc_per_node=2 train.py \
     --model DiT-XL/2 \
-    --batch_size 2
+    --batch_size 2 \
+    --num_classes 10
 ```
 
 We disable all speedup methods by default. Here are details of some key arguments for training:
 - `--nproc_per_node`: The GPU number you want to use for the current node.
 - `--plugin`: The booster plugin used by ColossalAI, `zero2` and `ddp` are supported. The default value is `zero2`. Recommend to enable `zero2`.
-- `--mixed_precision`: The data type for mixed precision training. The default value is `fp16`.
+- `--mixed_precision`: The data type for mixed precision training. The default value is `bf16`.
 - `--grad_checkpoint`: Whether enable the gradient checkpointing. This saves the memory cost during training process. The default value is `False`. Recommend to disable it when memory is enough.
-- `--enable_modulate_kernel`: Whether enable the modulate kernel optimization. This speeds up the training process. The default value is `False`. Recommend to enable it for GPU < H100.
 - `--enable_layernorm_kernel`: Whether enable the layernorm kernel optimization. This speeds up the training process. The default value is `False`. Recommend to enable it.
 - `--enable_flashattn`: Whether enable the FlashAttention. This speeds up the training process. The default value is `False`. Recommend to enable.
+- `--enable_modulate_kernel`: Whether enable the modulate kernel optimization. This speeds up the training process. The default value is `False`. This kernel will cause NaN under some circumstances. So we recommend to disable it for now.
 - `--sequence_parallel_size`: The sequence parallelism size. Will enable sequence parallelism when setting a value > 1. The default value is 1. Recommend to disable it if memory is enough.
+- `--load`: Load previous saved checkpoint dir and continue training.
+- `--num_classes`: Label class number. Should be 10 for CIFAR10 and 1000 for ImageNet. Only used for label-to-image generation.
+
 
 For more details on the configuration of the training process, please visit our code.
 
 <b>Multi-Node Training.</b>
 
-To train OpenDiT on mutiple nodes, you can use the following command:
+To train OpenDiT on multiple nodes, you can use the following command:
 
 ```
 colossalai run --nproc_per_node 8 --hostfile hostfile train.py \
     --model DiT-XL/2 \
-    --batch_size 2
+    --batch_size 2 \
+    --num_classes 10
 ```
 
 And you need to create `hostfile` under the current dir. It should contain all IP address of your nodes and you need to make sure all nodes can be connected without password by ssh. An example of hostfile:
@@ -133,18 +143,28 @@ And you need to create `hostfile` under the current dir. It should contain all I
 # Use script
 bash sample_img.sh
 # Use command line
-python sample.py --model DiT-XL/2 --image_size 256 --ckpt ./model.pt
+python sample.py \
+    --model DiT-XL/2 \
+    --image_size 256 \
+    --num_classes 10 \
+    --ckpt ckpt_path
 ```
+Here are details of some addtional key arguments for inference:
+- `--ckpt`: The weight of ema model `ema.pt`. To check your training progress, it can also be our saved base model `epochXX-global_stepXX/model`, it will produce better results than ema in early training stage.
+- `--num_classes`: Label class number. Should be 10 for CIFAR10, and 1000 for ImageNet (including official and our checkpoint).
 
 ### Video
-<b>Training.</b> Our video training pipeline is a faithful implementation, and we encourage you to explore your own strategies using OpenDiT. You can train the video DiT model by executing the following command:
+<b>Training.</b> We current support `VDiT` and `Latte` for video generation. VDiT adopts DiT structure and use video as inputs data. Latte further use more efficient spatial & temporal blocks based on VDiT (not exactly align with origin [Latte](https://github.com/Vchitect/Latte)).
+
+Our video training pipeline is a faithful implementation, and we encourage you to explore your own strategies using OpenDiT. You can train the video DiT model by executing the following command:
 
 ```shell
 # train with scipt
 bash train_video.sh
 # train with command line
+# model can also be Latte-XL/1x2x2
 torchrun --standalone --nproc_per_node=2 train.py \
-    --model vDiT-XL/222 \
+    --model VDiT-XL/1x2x2 \
     --use_video \
     --data_path ./videos/demo.csv \
     --batch_size 1 \
@@ -153,7 +173,7 @@ torchrun --standalone --nproc_per_node=2 train.py \
     --frame_interval 3
 
 # preprocess
-# our code read video from csv as the demo shows
+# our code read video from csv using our toy data
 # we provide a code to transfer ucf101 to csv format
 python preprocess.py
 ```
@@ -166,14 +186,17 @@ This script shares the same speedup methods as we have shown in the image traini
 # Use script
 bash sample_video.sh
 # Use command line
+# model can also be Latte-XL/1x2x2
 python sample.py \
-    --model vDiT-XL/222 \
+    --model VDiT-XL/1x2x2 \
     --use_video \
     --ckpt ckpt_path \
     --num_frames 16 \
     --image_size 256 \
     --frame_interval 3
 ```
+
+Inference tips: 1) EMA model requires quite long time to converge and produce meaningful results. So you can sample base model (`--ckpt /epochXX-global_stepXX/model`) instead of ema model (`--ckpt /epochXX-global_stepXX/ema.pt`) to check your training process. But ema model should be your final result. 2) Modify the text condition in `sample.py` which aligns with your datasets helps to produce better results in the early stage of training.
 
 ## FastSeq
 
@@ -210,7 +233,8 @@ torchrun --standalone --nproc_per_node=8 train.py \
     --batch_size 180 \
     --enable_layernorm_kernel \
     --enable_flashattn \
-    --mixed_precision fp16
+    --mixed_precision bf16 \
+    --num_classes 1000
 ```
