+++ codes/OpenDiT/opendit/models/dit.py
@@ -9,487 +9,24 @@
 # MAE:   https://github.com/facebookresearch/mae/blob/main/models_mae.py
 # --------------------------------------------------------
 
-import math
 from typing import Optional
 
 import numpy as np
 import torch
 import torch.distributed as dist
 import torch.nn as nn
-import torch.nn.functional as F
 import torch.utils.checkpoint
 from einops import rearrange
-from timm.models.vision_transformer import Mlp, PatchEmbed
+from timm.models.vision_transformer import PatchEmbed
 from torch.distributed import ProcessGroup
-from torch.jit import Final
 
-from opendit.models.clip import TextEmbedder
-from opendit.utils.operation import AllGather, AsyncAllGatherForTwo, all_to_all_comm, gather_forward_split_backward
-
-
-def get_layernorm(hidden_size: torch.Tensor, eps: float, affine: bool, use_kernel: bool):
-    if use_kernel:
-        try:
-            from apex.normalization import FusedLayerNorm
-
-            return FusedLayerNorm(hidden_size, elementwise_affine=affine, eps=eps)
-        except ImportError:
-            raise RuntimeError("FusedLayerNorm not available. Please install apex.")
-    else:
-        return nn.LayerNorm(hidden_size, eps, elementwise_affine=affine)
-
-
-def modulate(norm_func, x, shift, scale, use_kernel=False):
-    # Suppose x is (N, T, D), shift is (N, D), scale is (N, D)
-    dtype = x.dtype
-    x = norm_func(x.to(torch.float32)).to(dtype)
-    if use_kernel:
-        try:
-            from opendit.kernels.fused_modulate import fused_modulate
-
-            x = fused_modulate(x, scale, shift)
-        except ImportError:
-            raise RuntimeError("FusedModulate kernel not available. Please install triton.")
-    else:
-        x = x * (scale.unsqueeze(1) + 1) + shift.unsqueeze(1)
-    x = x.to(dtype)
-
-    return x
-
-
-def get_3d_sincos_pos_embed(embed_dim, grid_size, t_size, cls_token=False):
-    """
-    grid_size: int of the grid height and width
-    t_size: int of the temporal size
-    return:
-    pos_embed: [t_size*grid_size*grid_size, embed_dim] or [1+t_size*grid_size*grid_size, embed_dim] (w/ or w/o cls_token)
-    """
-    assert embed_dim % 4 == 0
-    embed_dim_spatial = embed_dim // 4 * 3
-    embed_dim_temporal = embed_dim // 4
-
-    # spatial
-    grid_h = np.arange(grid_size, dtype=np.float32)
-    grid_w = np.arange(grid_size, dtype=np.float32)
-    grid = np.meshgrid(grid_w, grid_h)  # here w goes first
-    grid = np.stack(grid, axis=0)
-
-    grid = grid.reshape([2, 1, grid_size, grid_size])
-    pos_embed_spatial = get_2d_sincos_pos_embed_from_grid(embed_dim_spatial, grid)
-
-    # temporal
-    grid_t = np.arange(t_size, dtype=np.float32)
-    pos_embed_temporal = get_1d_sincos_pos_embed_from_grid(embed_dim_temporal, grid_t)
-
-    # concate: [T, H, W] order
-    pos_embed_temporal = pos_embed_temporal[:, np.newaxis, :]
-    pos_embed_temporal = np.repeat(pos_embed_temporal, grid_size**2, axis=1)  # [T, H*W, D // 4]
-    pos_embed_spatial = pos_embed_spatial[np.newaxis, :, :]
-    pos_embed_spatial = np.repeat(pos_embed_spatial, t_size, axis=0)  # [T, H*W, D // 4 * 3]
-
-    pos_embed = np.concatenate([pos_embed_temporal, pos_embed_spatial], axis=-1)
-    pos_embed = pos_embed.reshape([-1, embed_dim])  # [T*H*W, D]
-
-    if cls_token:
-        pos_embed = np.concatenate([np.zeros([1, embed_dim]), pos_embed], axis=0)
-    return pos_embed
-
-
-#################################################################################
-#               Embedding Layers for Timesteps and Class Labels                 #
-#################################################################################
-
-
-class PatchEmbed3D(nn.Module):
-    """Video to Patch Embedding.
-
-    Args:
-        patch_size (int): Patch token size. Default: (2,4,4).
-        in_chans (int): Number of input video channels. Default: 3.
-        embed_dim (int): Number of linear projection output channels. Default: 96.
-        norm_layer (nn.Module, optional): Normalization layer. Default: None
-    """
-
-    def __init__(
-        self,
-        patch_size=(2, 4, 4),
-        in_chans=3,
-        embed_dim=96,
-        norm_layer=None,
-        flatten=True,
-    ):
-        super().__init__()
-        self.patch_size = patch_size
-        self.flatten = flatten
-
-        self.in_chans = in_chans
-        self.embed_dim = embed_dim
-
-        self.proj = nn.Conv3d(in_chans, embed_dim, kernel_size=patch_size, stride=patch_size)
-        if norm_layer is not None:
-            self.norm = norm_layer(embed_dim)
-        else:
-            self.norm = None
-
-    def forward(self, x):
-        """Forward function."""
-        # padding
-        _, _, D, H, W = x.size()
-        if W % self.patch_size[2] != 0:
-            x = F.pad(x, (0, self.patch_size[2] - W % self.patch_size[2]))
-        if H % self.patch_size[1] != 0:
-            x = F.pad(x, (0, 0, 0, self.patch_size[1] - H % self.patch_size[1]))
-        if D % self.patch_size[0] != 0:
-            x = F.pad(x, (0, 0, 0, 0, 0, self.patch_size[0] - D % self.patch_size[0]))
-
-        x = self.proj(x)  # B C D Wh Ww
-        if self.norm is not None:
-            D, Wh, Ww = x.size(2), x.size(3), x.size(4)
-            x = x.flatten(2).transpose(1, 2)
-            x = self.norm(x)
-            x = x.transpose(1, 2).view(-1, self.embed_dim, D, Wh, Ww)
-        if self.flatten:
-            x = x.flatten(2).transpose(1, 2)  # BCTHW -> BNC
-        return x
-
-
-class TimestepEmbedder(nn.Module):
-    """
-    Embeds scalar timesteps into vector representations.
-    """
-
-    def __init__(self, hidden_size, frequency_embedding_size=256):
-        super().__init__()
-        self.mlp = nn.Sequential(
-            nn.Linear(frequency_embedding_size, hidden_size, bias=True),
-            nn.SiLU(),
-            nn.Linear(hidden_size, hidden_size, bias=True),
-        )
-        self.frequency_embedding_size = frequency_embedding_size
-
-    @staticmethod
-    def timestep_embedding(t, dim, max_period=10000):
-        """
-        Create sinusoidal timestep embeddings.
-        :param t: a 1-D Tensor of N indices, one per batch element.
-                          These may be fractional.
-        :param dim: the dimension of the output.
-        :param max_period: controls the minimum frequency of the embeddings.
-        :return: an (N, D) Tensor of positional embeddings.
-        """
-        # https://github.com/openai/glide-text2im/blob/main/glide_text2im/nn.py
-        half = dim // 2
-        freqs = torch.exp(-math.log(max_period) * torch.arange(start=0, end=half, dtype=torch.float32) / half).to(
-            device=t.device
-        )
-        args = t[:, None].float() * freqs[None]
-        embedding = torch.cat([torch.cos(args), torch.sin(args)], dim=-1)
-        if dim % 2:
-            embedding = torch.cat([embedding, torch.zeros_like(embedding[:, :1])], dim=-1)
-        return embedding
-
-    def forward(self, t, dtype):
-        t_freq = self.timestep_embedding(t, self.frequency_embedding_size)
-        if t_freq.dtype != dtype:
-            t_freq = t_freq.to(dtype)
-        t_emb = self.mlp(t_freq)
-        return t_emb
-
-
-class LabelEmbedder(nn.Module):
-    """
-    Embeds class labels into vector representations. Also handles label dropout for classifier-free guidance.
-    """
-
-    def __init__(self, num_classes, hidden_size, dropout_prob):
-        super().__init__()
-        use_cfg_embedding = dropout_prob > 0
-        self.embedding_table = nn.Embedding(num_classes + use_cfg_embedding, hidden_size)
-        self.num_classes = num_classes
-        self.dropout_prob = dropout_prob
-
-    def token_drop(self, labels, force_drop_ids=None):
-        """
-        Drops labels to enable classifier-free guidance.
-        """
-        if force_drop_ids is None:
-            drop_ids = torch.rand(labels.shape[0], device=labels.device) < self.dropout_prob
-        else:
-            drop_ids = force_drop_ids == 1
-        labels = torch.where(drop_ids, self.num_classes, labels)
-        return labels
-
-    def forward(self, labels, train, force_drop_ids=None):
-        use_dropout = self.dropout_prob > 0
-        if (train and use_dropout) or (force_drop_ids is not None):
-            labels = self.token_drop(labels, force_drop_ids)
-        embeddings = self.embedding_table(labels)
-        return embeddings
-
-
-#################################################################################
-#                                 Core DiT Model                                #
-#################################################################################
-
-
-class DistAttention(nn.Module):
-    fused_attn: Final[bool]
-
-    def __init__(
-        self,
-        dim: int,
-        num_heads: int = 8,
-        qkv_bias: bool = False,
-        qk_norm: bool = False,
-        attn_drop: float = 0.0,
-        proj_drop: float = 0.0,
-        norm_layer: nn.Module = nn.LayerNorm,
-        enable_flashattn: bool = False,
-        sequence_parallel_size: int = 1,
-        sequence_parallel_group: Optional[ProcessGroup] = None,
-        sequence_parallel_type: str = None,
-        sequence_parallel_overlap: bool = False,
-        sequence_parallel_overlap_size: int = 2,
-    ) -> None:
-        super().__init__()
-        assert dim % num_heads == 0, "dim should be divisible by num_heads"
-        self.dim = dim
-        self.num_heads = num_heads
-        self.head_dim = dim // num_heads
-        self.scale = self.head_dim**-0.5
-        self.qkv = nn.Linear(dim, dim * 3, bias=qkv_bias)
-        self.q_norm = norm_layer(self.head_dim) if qk_norm else nn.Identity()
-        self.k_norm = norm_layer(self.head_dim) if qk_norm else nn.Identity()
-        self.attn_drop = nn.Dropout(attn_drop)
-        self.proj = nn.Linear(dim, dim)
-        self.proj_drop = nn.Dropout(proj_drop)
-        self.enable_flashattn = enable_flashattn
-
-        # sequence parallel
-        self.sequence_parallel_type = sequence_parallel_type
-        self.sequence_parallel_size = sequence_parallel_size
-        if sequence_parallel_size == 1:
-            sequence_parallel_type = None
-        else:
-            assert sequence_parallel_type in [
-                "longseq",
-                "ulysses",
-            ], "sequence_parallel_type should be longseq or ulysses"
-        self.sequence_parallel_group = sequence_parallel_group
-        self.sequence_parallel_overlap = sequence_parallel_overlap
-        self.sequence_parallel_overlap_size = sequence_parallel_overlap_size
-        if self.sequence_parallel_size > 1:
-            self.sequence_parallel_rank = dist.get_rank(sequence_parallel_group)
-            self.sequence_parallel_param_slice = slice(
-                self.qkv.out_features // sequence_parallel_size * self.sequence_parallel_rank,
-                self.qkv.out_features // sequence_parallel_size * (self.sequence_parallel_rank + 1),
-            )
-
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        B, N, C = x.shape
-        total_N = N * self.sequence_parallel_size
-
-        if self.sequence_parallel_type == "longseq":
-            if self.sequence_parallel_overlap:
-                if self.sequence_parallel_size == 2:
-                    # (B, N / SP_SIZE, C) => (SP_SIZE * B, N / SP_SIZE, C)
-                    qkv = AsyncAllGatherForTwo.apply(
-                        x,
-                        self.qkv.weight[self.sequence_parallel_param_slice],
-                        self.qkv.bias[self.sequence_parallel_param_slice],
-                        self.sequence_parallel_rank,
-                        self.sequence_parallel_size,
-                        dist.group.WORLD,
-                    )  # (B, N, C / SP_SIZE)
-                else:
-                    raise NotImplementedError(
-                        "sequence_parallel_overlap is only supported for sequence_parallel_size=2"
-                    )
-            else:
-                # (B, N / SP_SIZE, C) => (SP_SIZE * B, N / SP_SIZE, C)
-                x = AllGather.apply(x)[0]
-                # (SP_SIZE, B, N / SP_SIZE, C) => (B, N, C)
-                x = rearrange(x, "sp b n c -> b (sp n) c")
-                qkv = F.linear(
-                    x,
-                    self.qkv.weight[self.sequence_parallel_param_slice],
-                    self.qkv.bias[self.sequence_parallel_param_slice],
-                )
-        else:
-            qkv = self.qkv(x)  # (B, N, C), N here is N_total // SP_SIZE
-
-        num_heads = (
-            self.num_heads if self.sequence_parallel_type is None else self.num_heads // self.sequence_parallel_size
-        )
-
-        if self.sequence_parallel_type == "ulysses":
-            q, k, v = qkv.split(self.head_dim * self.num_heads, dim=-1)
-            q = all_to_all_comm(q, self.sequence_parallel_group)
-            k = all_to_all_comm(k, self.sequence_parallel_group)
-            v = all_to_all_comm(v, self.sequence_parallel_group)
-
-            if self.enable_flashattn:
-                q = q.reshape(B, N * self.sequence_parallel_size, num_heads, self.head_dim).contiguous()
-                k = k.reshape(B, N * self.sequence_parallel_size, num_heads, self.head_dim).contiguous()
-                v = v.reshape(B, N * self.sequence_parallel_size, num_heads, self.head_dim).contiguous()
-            else:
-                q = (
-                    q.reshape(B, N * self.sequence_parallel_size, num_heads, self.head_dim)
-                    .permute(0, 2, 1, 3)
-                    .contiguous()
-                )
-                k = (
-                    k.reshape(B, N * self.sequence_parallel_size, num_heads, self.head_dim)
-                    .permute(0, 2, 1, 3)
-                    .contiguous()
-                )
-                v = (
-                    v.reshape(B, N * self.sequence_parallel_size, num_heads, self.head_dim)
-                    .permute(0, 2, 1, 3)
-                    .contiguous()
-                )
-
-        else:
-            if self.sequence_parallel_type == "longseq":
-                qkv_shape = (B, total_N, num_heads, 3, self.head_dim)
-                if self.enable_flashattn:
-                    qkv_permute_shape = (3, 0, 1, 2, 4)
-                else:
-                    qkv_permute_shape = (3, 0, 2, 1, 4)
-            else:
-                qkv_shape = (B, total_N, 3, num_heads, self.head_dim)
-                if self.enable_flashattn:
-                    qkv_permute_shape = (2, 0, 1, 3, 4)
-                else:
-                    qkv_permute_shape = (2, 0, 3, 1, 4)
-            qkv = qkv.view(qkv_shape).permute(qkv_permute_shape)
-            q, k, v = qkv.unbind(0)
-        q, k = self.q_norm(q), self.k_norm(k)
-        if self.enable_flashattn:
-            from flash_attn import flash_attn_func
-
-            x = flash_attn_func(
-                q,
-                k,
-                v,
-                dropout_p=self.attn_drop.p if self.training else 0.0,
-                softmax_scale=self.scale,
-            )
-        else:
-            dtype = q.dtype
-            q = q * self.scale
-            attn = q @ k.transpose(-2, -1)
-            # translate attn to float32
-            attn = attn.to(torch.float32)
-            attn = attn.softmax(dim=-1)
-            # cast back attn to original dtype
-            attn = attn.to(dtype)
-            attn = self.attn_drop(attn)
-            x = attn @ v
-
-        if self.sequence_parallel_type is None:
-            x_output_shape = (B, N, C)
-        else:
-            x_output_shape = (B, total_N, num_heads * self.head_dim)
-        if self.enable_flashattn:
-            x = x.reshape(x_output_shape)
-        else:
-            x = x.transpose(1, 2).reshape(x_output_shape)
-        if self.sequence_parallel_size > 1:
-            x = all_to_all_comm(x, self.sequence_parallel_group, scatter_dim=1, gather_dim=2)
-        x = self.proj(x)
-        x = self.proj_drop(x)
-
-        return x
-
-    # Rearrange the qkv projection (qkv ... qkv <-> q ... q k ... k v ... v)
-    def rearrange_fused_weight(self, layer: nn.Linear, flag="load"):
-        # check whether layer is an torch.nn.Linear layer
-        if not isinstance(layer, nn.Linear):
-            raise ValueError("Invalid layer type for fused qkv weight rearrange!")
-
-        with torch.no_grad():
-            if flag == "load":
-                layer.weight.data = rearrange(layer.weight.data, "(x NH H) D -> (NH x H) D", x=3, H=self.head_dim)
-                layer.bias.data = rearrange(layer.bias.data, "(x NH H) -> (NH x H)", x=3, H=self.head_dim)
-                assert layer.weight.data.is_contiguous()
-                assert layer.bias.data.is_contiguous()
-
-            elif flag == "save":
-                layer.weight.data = rearrange(layer.weight.data, "(NH x H) D -> (x NH H) D", x=3, H=self.head_dim)
-                layer.bias.data = rearrange(layer.bias.data, "(NH x H) -> (x NH H)", x=3, H=self.head_dim)
-                assert layer.weight.data.is_contiguous()
-                assert layer.bias.data.is_contiguous()
-            else:
-                raise ValueError("Invalid flag for fused qkv weight rearrange!")
-
-
-class DiTBlock(nn.Module):
-    """
-    A DiT block with adaptive layer norm zero (adaLN-Zero) conditioning.
-    """
-
-    def __init__(
-        self,
-        hidden_size,
-        num_heads,
-        mlp_ratio=4.0,
-        enable_flashattn=False,
-        sequence_parallel_size: int = 1,
-        sequence_parallel_group: Optional[ProcessGroup] = None,
-        sequence_parallel_type: str = None,
-        enable_layernorm_kernel=False,
-        enable_modulate_kernel=False,
-        **block_kwargs,
-    ):
-        super().__init__()
-        self.enable_modulate_kernel = enable_modulate_kernel
-        self.norm1 = get_layernorm(hidden_size, eps=1e-6, affine=False, use_kernel=enable_layernorm_kernel)
-        self.attn = DistAttention(
-            hidden_size,
-            num_heads=num_heads,
-            qkv_bias=True,
-            enable_flashattn=enable_flashattn,
-            sequence_parallel_size=sequence_parallel_size,
-            sequence_parallel_group=sequence_parallel_group,
-            sequence_parallel_type=sequence_parallel_type,
-            **block_kwargs,
-        )
-        self.norm2 = get_layernorm(hidden_size, eps=1e-6, affine=False, use_kernel=enable_layernorm_kernel)
-        mlp_hidden_dim = int(hidden_size * mlp_ratio)
-        approx_gelu = lambda: nn.GELU(approximate="tanh")
-        self.mlp = Mlp(in_features=hidden_size, hidden_features=mlp_hidden_dim, act_layer=approx_gelu, drop=0)
-        self.adaLN_modulation = nn.Sequential(nn.SiLU(), nn.Linear(hidden_size, 6 * hidden_size, bias=True))
-
-    def forward(self, x, c):
-        shift_msa, scale_msa, gate_msa, shift_mlp, scale_mlp, gate_mlp = self.adaLN_modulation(c).chunk(6, dim=1)
-        x = x + gate_msa.unsqueeze(1) * self.attn(
-            modulate(self.norm1, x, shift_msa, scale_msa, self.enable_modulate_kernel)
-        )
-        x = x + gate_mlp.unsqueeze(1) * self.mlp(
-            modulate(self.norm2, x, shift_mlp, scale_mlp, self.enable_modulate_kernel)
-        )
-        return x
-
-
-class FinalLayer(nn.Module):
-    """
-    The final layer of DiT.
-    """
-
-    def __init__(self, hidden_size, patch_size, out_channels, use_video):
-        super().__init__()
-        self.norm_final = nn.LayerNorm(hidden_size, elementwise_affine=False, eps=1e-6)
-        self.linear = nn.Linear(
-            hidden_size, patch_size * out_channels if use_video else patch_size * patch_size * out_channels, bias=True
-        )
-        self.adaLN_modulation = nn.Sequential(nn.SiLU(), nn.Linear(hidden_size, 2 * hidden_size, bias=True))
-
-    def forward(self, x, c):
-        shift, scale = self.adaLN_modulation(c).chunk(2, dim=1)
-        x = modulate(self.norm_final, x, shift, scale)
-        x = self.linear(x)
-        return x
+from opendit.embed.clip_text_emb import TextEmbedder
+from opendit.embed.label_emb import LabelEmbedder
+from opendit.embed.patch_emb import PatchEmbed3D
+from opendit.embed.pos_emb import get_1d_sincos_pos_embed, get_2d_sincos_pos_embed
+from opendit.embed.time_emb import TimestepEmbedder
+from opendit.modules.block import DiTBlock, FinalLayer
+from opendit.utils.operation import gather_forward_split_backward
 
 
 class DiT(nn.Module):
@@ -524,6 +61,7 @@ class DiT(nn.Module):
         self.use_video = use_video
         self.in_channels = in_channels
         self.out_channels = in_channels * 2 if learn_sigma else in_channels
+        self.hidden_size = hidden_size
         self.patch_size = patch_size
         self.input_size = input_size
         self.num_heads = num_heads
@@ -542,17 +80,22 @@ class DiT(nn.Module):
         # For video input, use PatchEmbed3D and TextEmbedder
         if self.use_video:
             self.x_embedder = PatchEmbed3D(patch_size, in_channels, embed_dim=hidden_size)
+            self.t_embedder = TimestepEmbedder(hidden_size)
             self.y_embedder = TextEmbedder(path=text_encoder, hidden_size=hidden_size)
             self.num_patches = np.prod([input_size[i] // patch_size[i] for i in range(3)])
         else:
             self.x_embedder = PatchEmbed(input_size, patch_size, in_channels, hidden_size, bias=True)
+            self.t_embedder = TimestepEmbedder(hidden_size)
             self.y_embedder = LabelEmbedder(num_classes, hidden_size, class_dropout_prob)
             self.num_patches = self.x_embedder.num_patches
-        self.t_embedder = TimestepEmbedder(hidden_size)
 
-        # Will use fixed sin-cos embedding:
-        # TODO: add 3d pos embedding
-        self.pos_embed = nn.Parameter(torch.zeros(1, self.num_patches, hidden_size), requires_grad=False)
+        if self.use_video:
+            self.num_temporal = input_size[0] // patch_size[0]
+            self.num_spatial = self.num_patches // self.num_temporal
+            self.register_buffer("pos_embed_spatial", self.get_spatial_pos_embed())
+            self.register_buffer("pos_embed_temporal", self.get_temporal_pos_embed())
+        else:
+            self.pos_embed = nn.Parameter(torch.zeros(1, self.num_patches, hidden_size), requires_grad=False)
 
         self.blocks = nn.ModuleList(
             [
@@ -571,7 +114,7 @@ class DiT(nn.Module):
             ]
         )
         self.final_layer = FinalLayer(
-            hidden_size, np.prod(patch_size) if use_video else patch_size, self.out_channels, self.use_video
+            hidden_size, np.prod(patch_size) if use_video else patch_size ** 2, self.out_channels
         )
         self.initialize_weights()
 
@@ -579,6 +122,22 @@ class DiT(nn.Module):
 
     def enable_gradient_checkpointing(self):
         self.gradient_checkpointing = True
+
+    def get_spatial_pos_embed(self):
+        pos_embed = get_2d_sincos_pos_embed(
+            self.hidden_size,
+            self.input_size[1] // self.patch_size[1],
+        )
+        pos_embed = torch.from_numpy(pos_embed).float().unsqueeze(0).requires_grad_(False)
+        return pos_embed
+
+    def get_temporal_pos_embed(self):
+        pos_embed = get_1d_sincos_pos_embed(
+            self.hidden_size,
+            self.input_size[0] // self.patch_size[0],
+        )
+        pos_embed = torch.from_numpy(pos_embed).float().unsqueeze(0).requires_grad_(False)
+        return pos_embed
 
     def initialize_weights(self):
         # Initialize transformer layers:
@@ -592,15 +151,9 @@ class DiT(nn.Module):
         self.apply(_basic_init)
 
         # Initialize (and freeze) pos_embed by sin-cos embedding:
-        if self.use_video:
-            pos_embed = get_3d_sincos_pos_embed(
-                self.pos_embed.shape[-1],
-                self.input_size[-1] // self.patch_size[-1],
-                self.input_size[0] // self.patch_size[0],
-            )
-        else:
+        if not self.use_video:
             pos_embed = get_2d_sincos_pos_embed(self.pos_embed.shape[-1], int(self.x_embedder.num_patches**0.5))
-        self.pos_embed.data.copy_(torch.from_numpy(pos_embed).float().unsqueeze(0))
+            self.pos_embed.data.copy_(torch.from_numpy(pos_embed).float().unsqueeze(0))
 
         # Initialize patch_embed like nn.Linear (instead of nn.Conv2d):
         w = self.x_embedder.proj.weight.data
@@ -668,7 +221,16 @@ class DiT(nn.Module):
         # origin inputs should be float32, cast to specified dtype
         x = x.to(self.dtype)
 
-        x = self.x_embedder(x) + self.pos_embed  # (N, T, D), where T = H * W / patch_size ** 2
+        if self.use_video:
+            x = self.x_embedder(x)
+            x = rearrange(x, "b (t s) d -> b t s d", t=self.num_temporal, s=self.num_spatial)
+            x = x + self.pos_embed_spatial
+            x = rearrange(x, "b t s d -> b s t d")
+            x = x + self.pos_embed_temporal
+            x = rearrange(x, "b s t d -> b (t s) d")
+        else:
+            x = self.x_embedder(x) + self.pos_embed  # (N, T, D), where T = H * W / patch_size ** 2
+
         t = self.t_embedder(t, dtype=x.dtype)  # (N, D)
         y = self.y_embedder(y, self.training)  # (N, D)
         c = t + y  # (N, D)
@@ -718,62 +280,6 @@ class DiT(nn.Module):
         for block in self.blocks:
             block.attn.rearrange_fused_weight(block.attn.qkv, flag)
         torch.cuda.empty_cache()
-
-
-#################################################################################
-#                   Sine/Cosine Positional Embedding Functions                  #
-#################################################################################
-# https://github.com/facebookresearch/mae/blob/main/util/pos_embed.py
-
-
-def get_2d_sincos_pos_embed(embed_dim, grid_size, cls_token=False, extra_tokens=0):
-    """
-    grid_size: int of the grid height and width
-    return:
-    pos_embed: [grid_size*grid_size, embed_dim] or [1+grid_size*grid_size, embed_dim] (w/ or w/o cls_token)
-    """
-    grid_h = np.arange(grid_size, dtype=np.float32)
-    grid_w = np.arange(grid_size, dtype=np.float32)
-    grid = np.meshgrid(grid_w, grid_h)  # here w goes first
-    grid = np.stack(grid, axis=0)
-
-    grid = grid.reshape([2, 1, grid_size, grid_size])
-    pos_embed = get_2d_sincos_pos_embed_from_grid(embed_dim, grid)
-    if cls_token and extra_tokens > 0:
-        pos_embed = np.concatenate([np.zeros([extra_tokens, embed_dim]), pos_embed], axis=0)
-    return pos_embed
-
-
-def get_2d_sincos_pos_embed_from_grid(embed_dim, grid):
-    assert embed_dim % 2 == 0
-
-    # use half of dimensions to encode grid_h
-    emb_h = get_1d_sincos_pos_embed_from_grid(embed_dim // 2, grid[0])  # (H*W, D/2)
-    emb_w = get_1d_sincos_pos_embed_from_grid(embed_dim // 2, grid[1])  # (H*W, D/2)
-
-    emb = np.concatenate([emb_h, emb_w], axis=1)  # (H*W, D)
-    return emb
-
-
-def get_1d_sincos_pos_embed_from_grid(embed_dim, pos):
-    """
-    embed_dim: output dimension for each position
-    pos: a list of positions to be encoded: size (M,)
-    out: (M, D)
-    """
-    assert embed_dim % 2 == 0
-    omega = np.arange(embed_dim // 2, dtype=np.float64)
-    omega /= embed_dim / 2.0
-    omega = 1.0 / 10000**omega  # (D/2,)
-
-    pos = pos.reshape(-1)  # (M,)
-    out = np.einsum("m,d->md", pos, omega)  # (M, D/2), outer product
-
-    emb_sin = np.sin(out)  # (M, D/2)
-    emb_cos = np.cos(out)  # (M, D/2)
-
-    emb = np.concatenate([emb_sin, emb_cos], axis=1)  # (M, D)
-    return emb
 
 
 #################################################################################
@@ -829,17 +335,30 @@ def DiT_S_8(**kwargs):
     return DiT(depth=12, hidden_size=384, patch_size=8, num_heads=6, **kwargs)
 
 
-def vDiT_XL_222(**kwargs):
+def VDiT_XL_1x2x2(**kwargs):
+    return DiT(
+        depth=28,
+        hidden_size=1152,
+        patch_size=(1, 2, 2),
+        num_heads=16,
+        use_video=True,
+        **kwargs,
+    )
+
+
+def VDiT_XL_2x2x2(**kwargs):
     return DiT(
         depth=28,
         hidden_size=1152,
         patch_size=(2, 2, 2),
         num_heads=16,
+        use_video=True,
         **kwargs,
     )
 
 
 DiT_models = {
+    # image model
     "DiT-XL/2": DiT_XL_2,
     "DiT-XL/4": DiT_XL_4,
     "DiT-XL/8": DiT_XL_8,
@@ -852,5 +371,7 @@ DiT_models = {
     "DiT-S/2": DiT_S_2,
     "DiT-S/4": DiT_S_4,
     "DiT-S/8": DiT_S_8,
-    "vDiT-XL/222": vDiT_XL_222,
+    # video model
+    "VDiT-XL/1x2x2": VDiT_XL_1x2x2,
+    "VDiT-XL/2x2x2": VDiT_XL_2x2x2,
 }
