+++ codes/OpenDiT/opendit/utils/train_utils.py
@@ -5,6 +5,7 @@ import torch.distributed as dist
 from colossalai.zero.low_level.low_level_optim import LowLevelZeroOptimizer
 
 from opendit.models.dit import DiT
+from opendit.models.latte import Latte
 
 
 def get_model_numel(model: torch.nn.Module) -> int:
@@ -38,7 +39,8 @@ def update_ema(
     """
     Step the EMA model towards the current model.
     """
-    model = model if isinstance(model, DiT) else model.module
+    if not (isinstance(model, DiT) or isinstance(model, Latte)):
+        model = model.module
     ema_params = OrderedDict(ema_model.named_parameters())
     model_params = OrderedDict(model.named_parameters())
