+++ codes/OpenDiT/sample.py
@@ -15,8 +15,9 @@ import torch
 from diffusers.models import AutoencoderKL
 from torchvision.utils import save_image
 
-from opendit.models.diffusion import create_diffusion
+from opendit.diffusion import create_diffusion
 from opendit.models.dit import DiT_models
+from opendit.models.latte import Latte_models
 from opendit.utils.download import find_model
 from opendit.vae.reconstruct import save_sample
 from opendit.vae.wrapper import AutoencoderKLWrapper
@@ -51,9 +52,19 @@ def main(args):
         input_size = args.image_size // 8
 
     dtype = torch.float32
+    if "DiT" in args.model:
+        if "VDiT" in args.model:
+            assert args.use_video, "VDiT model requires video data"
+        else:
+            assert not args.use_video, "DiT model requires image data"
+        model_class = DiT_models[args.model]
+    elif "Latte" in args.model:
+        assert args.use_video, "Latte model requires video data"
+        model_class = Latte_models[args.model]
+    else:
+        raise ValueError(f"Unknown model {args.model}")
     model = (
-        DiT_models[args.model](
-            use_video=args.use_video,
+        model_class(
             input_size=input_size,
             num_classes=args.num_classes,
             enable_flashattn=False,
@@ -64,25 +75,31 @@ def main(args):
         .to(device)
         .to(dtype)
     )
+
     # Auto-download a pre-trained model or load a custom DiT checkpoint from train.py:
-    ckpt_path = args.ckpt or f"DiT-XL-2-{args.image_size}x{args.image_size}.pt"
+    ckpt_path = args.ckpt
     state_dict = find_model(ckpt_path)
     model.load_state_dict(state_dict)
     model.eval()  # important!
     diffusion = create_diffusion(str(args.num_sampling_steps))
 
-    # Labels to condition the model with (feel free to change):
-    class_labels = [207, 360, 387, 974, 88, 979, 417, 279]
-
     # Create sampling noise:
-    n = len(class_labels)
     if args.use_video:
+        # Labels to condition the model with (feel free to change):
+        class_labels = ["Biking", "Cliff Diving", "Rock Climbing Indoor", "Punch", "TaiChi"]
+        n = len(class_labels)
         z = torch.randn(n, vae.out_channels, *input_size, device=device)
-        y = ["video test"] * n * 2
+        y = class_labels * 2
     else:
+        # Labels to condition the model with (feel free to change):
+        if args.num_classes == 1000:
+            class_labels = [207, 360, 387, 974, 88, 979, 417, 279]
+        else:
+            class_labels = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9]
+        n = len(class_labels)
         z = torch.randn(n, 4, input_size, input_size, device=device)
         y = torch.tensor(class_labels, device=device)
-        y_null = torch.tensor([1000] * n, device=device)
+        y_null = torch.tensor([0] * n, device=device)
         y = torch.cat([y, y_null], 0)
 
     # Setup classifier-free guidance:
@@ -106,7 +123,9 @@ def main(args):
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
-    parser.add_argument("--model", type=str, choices=list(DiT_models.keys()), default="DiT-XL/2")
+    parser.add_argument(
+        "--model", type=str, choices=list(DiT_models.keys()) + list(Latte_models.keys()), default="DiT-XL/2"
+    )
     parser.add_argument("--vae", type=str, choices=["ema", "mse"], default="ema")
     parser.add_argument("--image_size", type=int, choices=[256, 512], default=256)
     parser.add_argument("--num_classes", type=int, default=1000)
