+++ codes/OpenDiT/sample_img.sh
@@ -1,1 +1,5 @@
-python sample.py --model DiT-XL/2 --image_size 256 --ckpt ./pretrained/DiT-XL-2-256x256.pt
+python sample.py \
+    --model DiT-XL/2 \
+    --image_size 256 \
+    --num_classes 10 \
+    --ckpt ckpt_path
