+++ codes/OpenDiT/sample_video.sh
@@ -1,5 +1,5 @@
 python sample.py \
-    --model vDiT-XL/222 \
+    --model VDiT-XL/1x2x2 \
     --use_video \
     --ckpt ckpt_path \
     --num_frames 16 \
