+++ codes/OpenDiT/tests/test_fastseq_parallel.py
@@ -6,7 +6,7 @@ import torch.distributed as dist
 from colossalai.testing import parameterize, rerun_if_address_is_in_use, spawn
 from torch.testing import assert_close
 
-from opendit.models.dit import DistAttention
+from opendit.modules.attn import DistAttention
 
 WORKERS = 4
 DTYPE = torch.float16
